refactor code

+++ dts-resourcemanager/src/main/java/io/dts/mt/MtResourceManager.java
@@ -29,19 +29,19 @@ public class MtResourceManager extends BaseResourceManager {
   }
 
   @Override
-  public void branchCommit(String xid, long branchId, String key, String udata, byte commitMode,
+  public void branchCommit(String xid, long branchId, String key, String udata, int commitMode,
       String retrySql) throws DtsException {
 
   }
 
   @Override
-  public void branchRollback(String xid, long branchId, String key, String udata, byte commitMode)
+  public void branchRollback(String xid, long branchId, String key, String udata, int commitMode)
       throws DtsException {
 
   }
 
   @Override
-  public void branchRollback(String xid, long branchId, String key, String udata, byte commitMode,
+  public void branchRollback(String xid, long branchId, String key, String udata, int commitMode,
       int isDelKey) throws DtsException {
 
   }
