+++ dts-resourcemanager/src/main/java/io/dts/resourcemanager/ResourceManager.java
@@ -30,12 +30,12 @@ public interface ResourceManager {
   public void reportUdata(String xid, long branchId, String key, String udata, boolean delay)
       throws DtsException;
 
-  public void branchCommit(String xid, long branchId, String key, String udata, byte commitMode,
+  public void branchCommit(String xid, long branchId, String key, String udata, int commitMode,
       String retrySql) throws DtsException;
 
-  public void branchRollback(String xid, long branchId, String key, String udata, byte commitMode)
+  public void branchRollback(String xid, long branchId, String key, String udata, int commitMode)
       throws DtsException;
 
-  public void branchRollback(String xid, long branchId, String key, String udata, byte commitMode,
+  public void branchRollback(String xid, long branchId, String key, String udata, int commitMode,
       int isDelKey) throws DtsException;
 }
